backend/pyproject.toml
[project]
name = "character-backend"
version = "0.1.0"
description = "FastAPI CRUD backend for character persistence."
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.30",
    "motor>=3.6",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/character_backend"]

[tool.ruff.lint]
select = ["E", "F", "I"]
